pyproject.toml
[project]
name = "vibe-trade-mcp"
version = "0.1.2"
description = "MCP server for trading strategy creation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "google-cloud-firestore>=2.14.0",
    "python-dotenv>=1.0.0",
    "jsonschema>=4.20.0",
]
authors = [
    {name = "Vibe Trade", email = "dev@vibe-trade.com"},
]
license = {text = "MIT"}
keywords = ["mcp", "trading", "strategy", "firestore"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# Map src/ subdirectories to vibe_trade_mcp/ namespace
packages = [
    "vibe_trade_mcp",
    "vibe_trade_mcp.api",
    "vibe_trade_mcp.db",
    "vibe_trade_mcp.models",
    "vibe_trade_mcp.scripts",
    "vibe_trade_mcp.tools",
]
package-dir = {"vibe_trade_mcp" = "src"}

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
]

[project.scripts]
main = "vibe_trade_mcp.main:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/scripts/*",  # Scripts don't need test coverage
    "*/db/firestore_client.py",  # Not used yet (kept for future Firestore features)
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
fail_under = 60
